+++ app/app.js
@@ -2,6 +2,10 @@ import Application from '@ember/application';
 import Resolver from 'ember-resolver';
 import loadInitializers from 'ember-load-initializers';
 import ENV from 'frontend-verenigingen-loket/config/environment';
+import { silenceEmptySyncRelationshipWarnings } from './utils/ember-data';
+
+silenceEmptySyncRelationshipWarnings();
+
 export default class App extends Application {
   modulePrefix = ENV.modulePrefix;
   podModulePrefix = ENV.podModulePrefix;