+++ app/components/recognition/form.js
@@ -5,17 +5,18 @@ import { errorValidation } from '../../validations/recognition-validation';
 import { tracked } from '@glimmer/tracking';
 import dateFormat from '../../helpers/date-format';
 import { task } from 'ember-concurrency';
+import dateYear from 'frontend-verenigingen-loket/helpers/date-year';
 
 export default class FormComponent extends Component {
-  items = ['College van burgemeester en schepenen', 'Andere'];
   @service currentAssociation;
   @service contactPoints;
   @service currentRecognition;
+  @service currentSession;
   @service store;
   @service router;
   @service toaster;
-  @service dateYear;
   @service file;
+
   @tracked validationErrors = {};
   @tracked legalResourceFile = this.currentRecognition.recognition
     ? this.currentRecognition.recognition.file
@@ -53,6 +54,26 @@ export default class FormComponent extends Component {
     }
     return null;
   }
+
+  constructor() {
+    super(...arguments);
+
+    this.getGoverningBodiesInTime.perform();
+  }
+
+  getGoverningBodiesInTime = task(async () => {
+    const currentAdministrativeUnitId = this.currentSession.group.id;
+
+    return await this.store.query('governing-body', {
+      'filter[governing-body][administrative-unit][:id:]':
+        currentAdministrativeUnitId,
+      include:
+        'governing-body.classification,governing-body.administrative-unit',
+      sort: ':no-case:governing-body.classification.pref-label,-start',
+      // Just a precaution, if we actually hit a large amount of items we should use a different solution
+      'page[size]': 100,
+    });
+  });
 
   formatDate(date) {
     const day = date.toLocaleDateString('nl-BE', { day: '2-digit' }),
@@ -118,14 +139,14 @@ export default class FormComponent extends Component {
   }
 
   async validateForm() {
-    const startTime = dateFormat.compute([
+    const startTime = dateFormat(
       this.currentRecognition.recognitionModel.startTime,
       'YYY-MM-DD',
-    ]);
-    const endTime = dateFormat.compute([
+    );
+    const endTime = dateFormat(
       this.currentRecognition.recognitionModel.endTime,
       'YYY-MM-DD',
-    ]);
+    );
 
     const [startDateExist, endDateExist] = await this.getRecognitionsInPeriod(
       startTime,
@@ -135,10 +156,7 @@ export default class FormComponent extends Component {
 
     const err = errorValidation.validate({
       ...this.currentRecognition.recognitionModel,
-      awardedBy:
-        this.items[0] === this.currentRecognition.selectedItem
-          ? this.currentRecognition.selectedItem
-          : this.currentRecognition.recognitionModel.awardedBy,
+      awardedBy: this.currentRecognition.selectedItem,
       file: isNew ? await this.legalResourceFile : null,
     });
     this.validationErrors = err.error
@@ -197,16 +215,14 @@ export default class FormComponent extends Component {
       legalResource: recognitionModel.legalResource,
       association: this.currentAssociation.association,
       validityPeriod: await this.updateOrGetPeriod(recognitionModel),
-      awardedBy: await this.getAwardedBy(),
+      awardedBy: this.currentRecognition.selectedItem,
       file: fileData || null,
     });
 
     try {
       await recognition.save();
-      const startYear = this.dateYear.getCurrentYear(
-        recognitionModel.startTime,
-      );
-      const endYear = this.dateYear.getCurrentYear(recognitionModel.endTime);
+      const startYear = dateYear(recognitionModel.startTime);
+      const endYear = dateYear(recognitionModel.endTime);
       this.notify(
         `De erkenning voor de periode ${startYear} - ${endYear} is aangemaakt.`,
         `Erkenning succesvol aangemaakt.`,
@@ -242,7 +258,7 @@ export default class FormComponent extends Component {
     const recognition = {
       dateDocument: recognitionModel.dateDocument,
       legalResource: recognitionModel.legalResource,
-      awardedBy: await this.getAwardedBy(),
+      awardedBy: this.currentRecognition.selectedItem,
       validityPeriod,
       file: (await fileData) || null,
     };
@@ -251,8 +267,8 @@ export default class FormComponent extends Component {
         ...recognition,
       });
       await this.currentRecognition.recognition.save();
-      const startYear = this.dateYear.getCurrentYear(validityPeriod.startTime);
-      const endYear = this.dateYear.getCurrentYear(validityPeriod.endTime);
+      const startYear = dateYear(validityPeriod.startTime);
+      const endYear = dateYear(validityPeriod.endTime);
       this.notify(
         `De erkenning voor de periode ${startYear} - ${endYear} is bijgewerkt.`,
         `Erkenning succesvol bijgewerkt.`,
@@ -316,26 +332,11 @@ export default class FormComponent extends Component {
   }
 
   @action
-  async getAwardedBy() {
-    let awardedByValue = this.items[0];
-    if (this.currentRecognition.selectedItem !== this.items[0]) {
-      awardedByValue = this.currentRecognition.recognitionModel.awardedBy;
-    }
-
-    const administrativeUnits = await this.store.query('administrative-unit', {
-      filter: {
-        ':exact:name': awardedByValue,
-      },
-    });
-    let awardedBy = administrativeUnits?.[0];
-    if (!awardedBy) {
-      awardedBy = this.store.createRecord('administrative-unit', {
-        name: awardedByValue,
-      });
-      await awardedBy.save();
-    }
-    return awardedBy;
+  handleAwardedByChange(governingBody) {
+    this.currentRecognition.selectedItem = governingBody;
+    this.clearFormError('awardedBy');
   }
+
   @action
   async handleFileChange(event) {
     this.legalResourceFile = event.target.files[0];