+++ app/controllers/associations.js
@@ -3,8 +3,11 @@ import { service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 import { timeout, task } from 'ember-concurrency';
 import { action } from '@ember/object';
+import { cell, resource, resourceFactory } from 'ember-resources';
 import ENV from 'frontend-verenigingen-loket/config/environment';
 import { associationsQuery } from '../services/query-builder';
+import { dedupeTracked } from '../utils/tracked-toolbox';
+
 const DEBOUNCE_MS = 500;
 
 export default class IndexController extends Controller {
@@ -24,7 +27,7 @@ export default class IndexController extends Controller {
   @tracked selectedActivities = [];
   @tracked status = '';
   @tracked selectedOrganizationStatus = '';
-  @tracked postalCodes = '';
+  @dedupeTracked postalCodes = '';
   @tracked types = '';
   @tracked selectedTypes = [];
   @tracked targetAudiences = '';
@@ -32,6 +35,7 @@ export default class IndexController extends Controller {
   @tracked end = '';
   @tracked start = '';
   @tracked ENVIRONMENT_NAME = ENV.environmentName;
+  PostalCodes = PostalCodes;
 
   queryParams = [
     'sort',
@@ -45,10 +49,6 @@ export default class IndexController extends Controller {
     'end',
     'start',
   ];
-
-  get selectedPostalCodes() {
-    return this.postalCodes ? this.postalCodes.split(',') : [];
-  }
 
   @action
   setActivities(selectedActivities) {
@@ -74,6 +74,10 @@ export default class IndexController extends Controller {
     this.postalCodes = selectedPostals
       .map((postal) => postal.postalCode)
       .join(',');
+
+    // We update the postalCodes value so the interface updates immediately instead of having to wait until the fetch finishes.
+    // TODO: There has to be a better pattern so the resource stays self-contained.
+    postalCodes.current = selectedPostals;
   }
 
   @action
@@ -299,3 +303,34 @@ export default class IndexController extends Controller {
     URL.revokeObjectURL(href);
   };
 }
+
+// We store the postalCodes cell in module scope so the previous data is retained between data fetches.
+// This prevents the select from temporarily showing an empty selection.
+const postalCodes = cell([]);
+
+// This fetches the postal code records based on the postalCodes query param
+function PostalCodes(store, postalCodesQP) {
+  return resource(() => {
+    (async () => {
+      if (postalCodesQP) {
+        const records = await Promise.all(
+          postalCodesQP.split(',').map(async (postalCode) => {
+            const result = await store.query('postal-code', {
+              'filter[postal-code]': postalCode,
+            });
+
+            return result.at(0);
+          }),
+        );
+
+        postalCodes.current = records;
+      } else {
+        await Promise.resolve();
+        postalCodes.current = [];
+      }
+    })();
+
+    return postalCodes;
+  });
+}
+resourceFactory(PostalCodes);