+++ app/routes/association/recognition/edit.js
@@ -42,7 +42,12 @@ export default class AssociationRecognitionEditRoute extends Route {
         id: recognitionId,
         ':has-no:status': true,
       },
-      include: ['awarded-by', 'validity-period', 'file'].join(','),
+      include: [
+        'awarded-by.governing-body.classification',
+        'awarded-by.governing-body.administrative-unit.classification',
+        'validity-period',
+        'file',
+      ].join(','),
     });
   });
 }