+++ app/routes/association/recognition/index.js
@@ -17,7 +17,12 @@ export default class AssociationRecognitionRoute extends Route {
 
   loadRecognition = task({ keepLatest: true }, async (id, params) => {
     return await this.store.query('recognition', {
-      include: ['awarded-by', 'validity-period', 'file'].join(','),
+      include: [
+        'awarded-by.governing-body.classification',
+        'awarded-by.governing-body.administrative-unit.classification',
+        'validity-period',
+        'file',
+      ].join(','),
       filter: {
         ':has-no:status': true,
         association: {