+++ app/routes/associations.js
@@ -43,6 +43,7 @@ export default class AssociationsRoute extends Route {
         params,
         PAGE_SIZE,
       );
+
       return {
         exportFile,
         associations,