+++ app/services/current-recognition.js
@@ -2,10 +2,10 @@ import Service, { inject as service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 
 export default class CurrentRecognitionService extends Service {
-  items = ['College van burgemeester en schepenen', 'Andere'];
   @service store;
+
   @tracked recognition = null;
-  @tracked selectedItem = this.items[0];
+  @tracked selectedItem;
   @tracked isLoading = false;
   @tracked generalError = '';
 
@@ -45,8 +45,6 @@ export default class CurrentRecognitionService extends Service {
       legalResource: null,
       file: null,
     };
-    this.awardedBy = null;
-    this.selectedItem = this.items[0];
     this.recognition = recognition;
 
     if (recognition) {
@@ -56,11 +54,8 @@ export default class CurrentRecognitionService extends Service {
       this.recognitionModel.endTime =
         await recognition.validityPeriod.get('endTime');
       this.recognitionModel.legalResource = recognition.legalResource ?? null;
-      this.recognitionModel.awardedBy = await recognition.awardedBy.get('name');
-      this.selectedItem =
-        this.recognitionModel.awardedBy === this.items[0]
-          ? this.items[0]
-          : this.items[1];
+      this.recognitionModel.awardedBy = await recognition.awardedBy;
+      this.selectedItem = await recognition.awardedBy;
     }
   }
 }